Move special printing macros to `PrettyPrinting`
